requirements.txt
python-decouple==3.6
requests==2.22.0
